NBSN95 V1.1.1

+++ NBSN95/Drivers/BSP/src/sht20.c
@@ -102,8 +102,8 @@ void sht20Data(void)
   user_main_printf("Humidity =%.2f %%rh",hum);
 	user_main_printf("tem =%.2f C",tem);	
 	
-	sensor.temSHT = tem*10;
-	sensor.humSHT = hum*10;
+	sensor.temSHT = (int)(tem*10);
+	sensor.humSHT = (int)(hum*10);
 }
 
 void MX_I2C1_Init(void)
