+++ NBSN95/Drivers/BSP/src/sht31.c
@@ -78,6 +78,6 @@ void sht31Data(void)
   user_main_printf("Humidity =%.2f %%rh",hum);
 	user_main_printf("tem =%.2f C",tem);	
 	
-	sensor.temSHT = tem*10;
-	sensor.humSHT = hum*10;
+	sensor.temSHT = (int)(tem*10);
+	sensor.humSHT = (int)(hum*10);
 }
